# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  handlers.CourseAccessEnvelope:
    properties:
      data:
        $ref: '#/definitions/handlers.CourseAccessResponse'
      success:
        example: true
        type: boolean
    type: object
  handlers.CourseAccessResponse:
    properties:
      can_access_content:
        type: boolean
      course_id:
        type: string
      enrollment:
        $ref: '#/definitions/handlers.EnrollmentResponse'
      is_platform_admin:
        type: boolean
      is_teacher:
        type: boolean
    type: object
  handlers.CourseEnvelope:
    properties:
      data:
        $ref: '#/definitions/handlers.CourseResponse'
      success:
        example: true
        type: boolean
    type: object
...
  handlers.CourseResponse:
    properties:
      category_id:
        type: string
      category_name:
        type: string
      cover_url:
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      enrollment_capacity:
        type: integer
      enrollment_open:
        type: boolean
      id:
        type: string
      language:
        type: string
      level:
        type: string
      organization_id:
        type: string
      price_cents:
        type: integer
      published_at:
        type: string
      slug:
        type: string
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      teacher_id:
        type: string
      teacher_name:
        type: string
      title:
        type: string
      updated_at:
        type: string
      waitlist_enabled:
        type: boolean
    type: object
...
  handlers.EnrollmentEnvelope:
    properties:
      data:
        $ref: '#/definitions/handlers.EnrollmentResponse'
      success:
        example: true
        type: boolean
    type: object
  handlers.EnrollmentListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.EnrollmentResponse'
        type: array
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        example: true
        type: boolean
    type: object
  handlers.EnrollmentResponse:
    properties:
      cancelled_at:
        type: string
      course_id:
        type: string
      course_slug:
        type: string
      course_title:
        type: string
      created_at:
        type: string
      enrolled_at:
        type: string
      id:
        type: string
      invite_code_id:
        type: string
      source:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_email:
        type: string
      user_full_name:
        type: string
      user_id:
        type: string
      waitlisted_at:
        type: string
    type: object
...
  handlers.InviteCodeEnvelope:
    properties:
      data:
        $ref: '#/definitions/handlers.InviteCodeResponse'
      success:
        example: true
        type: boolean
    type: object
  handlers.InviteCodeListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.InviteCodeResponse'
        type: array
      success:
        example: true
        type: boolean
    type: object
  handlers.InviteCodeResponse:
    properties:
      code:
        type: string
      course_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      expires_at:
        type: string
      id:
        type: string
      max_uses:
        type: integer
      revoked_at:
        type: string
      updated_at:
        type: string
      uses_count:
        type: integer
    type: object
...
  handlers.MessageData:
    properties:
      message:
        type: string
      reset_token:
        type: string
    type: object
  handlers.MessageResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.MessageData'
      success:
        example: true
        type: boolean
    type: object
...
  handlers.createInviteCodeRequest:
    properties:
      code:
        type: string
      expires_at:
        type: string
      max_uses:
        type: integer
    type: object
...
  handlers.enrollRequest:
    properties:
      invite_code:
        type: string
    type: object
  handlers.enrollmentSettingsRequest:
    properties:
      enrollment_capacity:
        type: integer
      enrollment_open:
        type: boolean
      waitlist_enabled:
        type: boolean
    type: object
...
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@asnakech.com
    name: API Support
    url: http://www.asnakech.com/support
  description: |-
    Backend API for the Asnakech online education platform.
    All JSON responses use a shared envelope: `{ success, data, error, meta }`.
    Every response includes an `X-Request-ID` header.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://www.asnakech.com/terms
  title: Asnakech School API
  version: 0.1.0
paths:
...
  /api/v1/courses/{id}/access:
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CourseAccessEnvelope'
      security:
      - BearerAuth: []
      summary: Check course content access
      tags:
      - enrollments
...
  /api/v1/courses/{id}/enroll:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EnrollmentEnvelope'
      security:
      - BearerAuth: []
      summary: Unenroll from a course
      tags:
      - enrollments
    post:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Optional invite code
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.enrollRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.EnrollmentEnvelope'
      security:
      - BearerAuth: []
      summary: Enroll in a course
      tags:
      - enrollments
  /api/v1/courses/{id}/enrollment-settings:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.enrollmentSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CourseEnvelope'
      security:
      - BearerAuth: []
      summary: Update course enrollment settings
      tags:
      - enrollments
  /api/v1/courses/{id}/enrollments:
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      - description: active|waitlisted|cancelled
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EnrollmentListEnvelope'
      security:
      - BearerAuth: []
      summary: List enrollments for a course
      tags:
      - enrollments
  /api/v1/courses/{id}/invite-codes:
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.InviteCodeListEnvelope'
      security:
      - BearerAuth: []
      summary: List enrollment invite codes
      tags:
      - enrollments
    post:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Invite
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.createInviteCodeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.InviteCodeEnvelope'
      security:
      - BearerAuth: []
      summary: Create enrollment invite code
      tags:
      - enrollments
  /api/v1/courses/{id}/invite-codes/{codeId}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Invite code ID
        in: path
        name: codeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      security:
      - BearerAuth: []
      summary: Revoke enrollment invite code
      tags:
      - enrollments
...
  /api/v1/me/enrollments:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      - description: active|waitlisted|cancelled
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EnrollmentListEnvelope'
      security:
      - BearerAuth: []
      summary: List my enrollments
      tags:
      - enrollments
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"